feat: rename and structure files
